refactor: Move authentication and messaging logic to hooks for better organization
