+++ context/SocketContextProvider.tsx
@@ -1,4 +1,4 @@
-import { getStoredUser } from "@/lib/auth";
+import { getStoredUser } from "@/hooks/auth";
 import React, { createContext, useContext, useEffect, useState } from "react";
 import { io, Socket } from "socket.io-client";
